thermalctld: Ignore exception when deleting chassisdb entry fails
After sonic-mgmt tests, it does a config consistency check and if the
config consistency check fails, it runs recover_chassis(). That
function simultaneously runs config_reload on the supervisor and all
linecards (in parallel). This means that the supervisor networking can
be in the process of restarting when thermalctld exits and goes
through the cleanup process of deleting entries from chassisdb.